+++ package.json
@@ -33,6 +33,9 @@
   "devDependencies": {
     "@nhs-llc/gulp-jsforce-deploy": "^2.2.0",
     "@types/istanbul-lib-coverage": "^2.0.3",
+    "@types/uuid": "^8.3.0",
+    "@types/vinyl": "^2.0.4",
+    "@types/xml2js": "^0.4.8",
     "fs-extra": "^8.1.0",
     "gulp": "^4.0.2",
     "gulp-shell": "^0.7.1",
@@ -47,7 +50,7 @@
     "istanbul-lib-report": "^3.0.0",
     "istanbul-reports": "^3.0.0",
     "jest-stare": "^1.26.0",
-    "uuid": "^3.4.0",
+    "uuid": "^8.3.2",
     "vinyl": "^2.2.0",
     "xml2js": "^0.4.23"
   }
